+++ lua/plugins/config/focus_config.lua
@@ -4,7 +4,14 @@ return function(use)
 		config = function()
 			require('focus').setup({
 				signcolumn = false, -- disable sign column only on main window
-				excluded_filetypes = { 'calltree' },
+				excluded_filetypes = {
+					'calltree',
+					'dapui_scopes',
+					'dapui_breakpoints',
+					'dapui_stacks',
+					'dapui_watches',
+					'dap-repl',
+				},
 			})
 		end,
 		event = 'VimEnter',